% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonex.R
\name{phonex}
\alias{phonex}
\title{phonex}
\usage{
phonex(txt, char_to_remove = " -.+*/,:;_", convert_to_num = TRUE)
}
\arguments{
\item{txt}{a character string}

\item{char_to_remove}{characters to remove from the initial text (defaults: ' -.+*/,:;_')}

\item{convert_to_num}{if TRUE, converts the processed text to a base22 number (as in the original algorithm). if FALSE, keeps the last output before this conversion (defaults: TRUE)}
}
\description{
Transforms text according to the french adapted phonex algorithm proposed by Frédéric BROUARD. This work is based on the python implementation provided by  Christian Pennaforte
}
