% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundex2.R
\name{soundex2}
\alias{soundex2}
\title{soundex2}
\usage{
soundex2(txt)
}
\arguments{
\item{txt}{character string}
}
\description{
Transforms text according to the french adapted soundex2 algorithm proposed by Florent Carlier
}
